use truncated salted sha for token
initially thought it would be simpler and safer to use a string and
increment it but you need to fetch the previous record's token first
which can cause a race condition. Using a sha1 with a random salt and
truncating should yeild sufficient randomnness and not be prone to
collisions.